Move unit_arg to its own module
